requirements.txt
django
djangorestframework
environs
python-dotenv
langchain-community
unstructured
chromadb
tiktoken
python-magic
pytest-django